implement the first services (not tested yet)

+++ src/main/java/shop/plumeria/plummity/dto/UserDTO.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @Setter
 public class UserDTO {
     private String identifier;
-    private int points;
+    private long points;
     private boolean isVeteran;
 }
